chore: organize clis and tools
