+++ cmd/argocd/argocd-server/commands/common.go
