+++ tools/argocd-util/commands/secrets_redactor_test.go
