+++ tools/argocd-util/commands/settings_test.go
