+++ tools/argocd-util/commands/testdata/test_clusterrole.yaml
