+++ tools/argocd/commands/cert.go
