+++ tools/argocd/commands/common.go
