+++ tools/argocd/commands/common_test.go
