+++ tools/argocd/commands/completion.go
