+++ tools/argocd/commands/login_test.go
