+++ tools/argocd/commands/logout.go
