+++ tools/argocd/commands/logout_test.go
