+++ tools/argocd/commands/projectwindows.go
