+++ tools/argocd/commands/version.go
