+++ tools/argocd/util/common.go
