+++ tools/argocd/util/testdata/test.cert.pem
